+++ Assets/Scenes/BootScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 243f5569a519738418f8906f8363b36f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
