+++ Assets/Scenes/HomeScene.unity
@@ -406,7 +406,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e46f6bbacb6f7044ebbc775e45094c6b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  photonManager: {fileID: 1824652881}
 --- !u!1 &914843357
 GameObject:
   m_ObjectHideFlags: 0
@@ -840,18 +839,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1486548420}
   m_CullTransparentMesh: 1
---- !u!114 &1824652881 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 2140714015081443516, guid: b10a99a68833cd7429fbf11fca6a8819,
-    type: 3}
-  m_PrefabInstance: {fileID: 6786331071929012364}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1f91befd907511d489a43d3f5c052635, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1863719530
 GameObject:
   m_ObjectHideFlags: 0
@@ -1055,74 +1042,6 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &6786331071929012364
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 1585072811369884453, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_Name
-      value: AkimichiPhotonManager
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6478752164039561029, guid: b10a99a68833cd7429fbf11fca6a8819,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: b10a99a68833cd7429fbf11fca6a8819, type: 3}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
@@ -1131,4 +1050,3 @@ SceneRoots:
   - {fileID: 377157533}
   - {fileID: 2129832351}
   - {fileID: 504025116}
-  - {fileID: 6786331071929012364}
