+++ Assets/Scripts/Scenes/Boot.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f37f7a75be424245a13aab339f4fcad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
