+++ Assets/Scripts/Scenes/Boot/BootScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4985e847fb61a544d9b92b0dca11cf96
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
