+++ Assets/Scripts/Scenes/Home/HomeScene.cs
@@ -4,17 +4,11 @@ namespace Akimichi
 {
     public class HomeScene : MonoBehaviour
     {
-        [SerializeField]
-        private AkimichiPhotonManager photonManager = null;
+        
 
         private void Awake()
         {
             NetworkManager.Instance().Disconnect();
-            TransitionManager.Instance().AddScene(SceneConst.Home);
-            if (TransitionManager.Instance().IsFirstTransedScene(SceneConst.Home))
-            {
-                DontDestroyOnLoad(photonManager);
-            }
         }
 
         private void Start()
