+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,9 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
+    path: Assets/Scenes/BootScene.unity
+    guid: 243f5569a519738418f8906f8363b36f
+  - enabled: 1
     path: Assets/Scenes/TitleScene.unity
     guid: 53edc3fb701b66d4eb1109c52a708ad5
   - enabled: 1
